Complete refactoring of aws_auth_providers
