+++ aws/rust-runtime/aws-config/src/default_provider.rs
@@ -22,12 +22,240 @@ pub mod region {
 /// Default credentials provider chain
 pub mod credentials {
     use crate::environment::credentials::EnvironmentVariableCredentialsProvider;
-    use aws_types::credentials::ProvideCredentials;
+    use crate::meta::credentials::{CredentialsProviderChain, LazyCachingCredentialsProvider};
+    use aws_types::credentials::{future, ProvideCredentials};
+    use aws_types::os_shim_internal::{Env, Fs};
+    use aws_types::region::Region;
+    use smithy_async::rt::sleep::AsyncSleep;
+    use smithy_client::erase::DynConnector;
+    use std::borrow::Cow;
 
-    /// Default Region Provider chain
+    #[cfg(any(feature = "rustls", feature = "native-tls"))]
+    /// Default Credentials Provider chain
     ///
-    /// This provider will load region from environment variables.
-    pub fn default_provider() -> impl ProvideCredentials {
-        EnvironmentVariableCredentialsProvider::new()
+    /// The region from the default region provider will be used
+    pub async fn default_provider() -> impl ProvideCredentials {
+        use crate::meta::region::ProvideRegion;
+        let region = super::region::default_provider().region().await;
+        let mut builder = DefaultCredentialsChain::builder();
+        builder.set_region(region);
+        builder.build()
+    }
+
+    /// Default AWS Credential Provider Chain
+    ///
+    /// Resolution order:
+    /// 1. Environment variables: [`EnvironmentVariableCredentialsProvider`](crate::environment::EnvironmentVariableCredentialsProvider)
+    /// 2. Shared config (`~/.aws/config`, `~/.aws/credentials`): [`SharedConfigCredentialsProvider`](crate::profile::ProfileFileCredentialsProvider)
+    ///
+    /// The outer provider is wrapped in a refreshing cache.
+    ///
+    /// More providers are a work in progress.
+    ///
+    /// ## Example:
+    /// Create a default chain with a custom region:
+    /// ```rust
+    /// use aws_types::region::Region;
+    /// use aws_config::default_provider::credentials::DefaultCredentialsChain;
+    /// let credentials_provider = DefaultCredentialsChain::builder()
+    ///     .region(Region::new("us-west-1"))
+    ///     .build();
+    /// ```
+    ///
+    /// Create a default chain with no overrides:
+    /// ```rust
+    /// use aws_config::default_provider::credentials::DefaultCredentialsChain;
+    /// let credentials_provider = DefaultCredentialsChain::builder().build();
+    /// ```
+    #[derive(Debug)]
+    pub struct DefaultCredentialsChain(LazyCachingCredentialsProvider);
+
+    impl DefaultCredentialsChain {
+        /// Builder for `DefaultCredentialsChain`
+        pub fn builder() -> Builder {
+            Builder::default()
+        }
+    }
+
+    impl ProvideCredentials for DefaultCredentialsChain {
+        fn provide_credentials<'a>(&'a self) -> future::ProvideCredentials<'a>
+        where
+            Self: 'a,
+        {
+            self.0.provide_credentials()
+        }
+    }
+
+    /// Builder for [`DefaultCredentialsChain`](DefaultCredentialsChain)
+    #[derive(Default)]
+    pub struct Builder {
+        profile_file_builder: crate::profile::credentials::Builder,
+        web_identity_builder: crate::web_identity_token::Builder,
+        credential_cache: crate::meta::credentials::lazy_caching::Builder,
+        env: Option<Env>,
+    }
+
+    impl Builder {
+        /// Sets the region used when making requests to AWS services
+        ///
+        /// When unset, the default region resolver chain will be used.
+        pub fn region(mut self, region: Region) -> Self {
+            self.set_region(Some(region));
+            self
+        }
+
+        /// Sets the region used when making requests to AWS services
+        ///
+        /// When unset, the default region resolver chain will be used.
+        pub fn set_region(&mut self, region: Option<Region>) -> &mut Self {
+            self.profile_file_builder.set_region(region.clone());
+            self.web_identity_builder.set_region(region);
+            self
+        }
+
+        /// Override the HTTPS connector used for this provider
+        ///
+        /// If a connector other than Hyper is used or if the Tokio/Hyper features have been disabled
+        /// this method MUST be used to specify a custom connector.
+        pub fn connector(mut self, connector: DynConnector) -> Self {
+            self.profile_file_builder
+                .set_connector(Some(connector.clone()));
+            self.web_identity_builder.set_connector(Some(connector));
+            self
+        }
+
+        /// Override the sleep implementation used for this provider
+        ///
+        /// By default, Tokio will be used to support async sleep during credentials for timeouts
+        /// and reloading credentials. If the tokio default feature has been disabled, a custom
+        /// sleep implementation must be provided.
+        pub fn sleep(mut self, sleep: impl AsyncSleep + 'static) -> Self {
+            self.credential_cache = self.credential_cache.sleep(sleep);
+            self
+        }
+
+        /// Add an additional credential source for the ProfileProvider
+        ///
+        /// Assume role profiles may specify named credential sources:
+        /// ```ini
+        /// [default]
+        /// role_arn = arn:aws:iam::123456789:role/RoleA
+        /// credential_source = MyCustomProvider
+        /// ```
+        ///
+        /// Typically, these are built-in providers like `Environment`, however, custom sources may
+        /// also be used.
+        ///
+        /// See [`with_custom_provider`](crate::profile::credentials::Builder::with_custom_provider)
+        pub fn with_custom_credential_source(
+            mut self,
+            name: impl Into<Cow<'static, str>>,
+            provider: impl ProvideCredentials + 'static,
+        ) -> Self {
+            self.profile_file_builder = self
+                .profile_file_builder
+                .with_custom_provider(name, provider);
+            self
+        }
+
+        #[doc(hidden)]
+        /// Override the filesystem used for this provider
+        ///
+        /// This method exists to test credential providers
+        pub fn fs(mut self, fs: Fs) -> Self {
+            self.profile_file_builder.set_fs(fs.clone());
+            self.web_identity_builder.set_fs(fs);
+            self
+        }
+
+        #[doc(hidden)]
+        /// Override the environment used for this provider
+        ///
+        /// This method exists to test credential providers
+        pub fn env(mut self, env: Env) -> Self {
+            self.env = Some(env.clone());
+            self.profile_file_builder.set_env(env.clone());
+            self.web_identity_builder.set_env(env);
+            self
+        }
+
+        /// Creates a `DefaultCredentialsChain`
+        ///
+        /// ## Panics
+        /// This function will panic if no connector has been set and neither `rustls` and `native-tls`
+        /// features have both been disabled.
+        pub fn build(self) -> DefaultCredentialsChain {
+            let profile_provider = self.profile_file_builder.build();
+            let env_provider =
+                EnvironmentVariableCredentialsProvider::new_with_env(self.env.unwrap_or_default());
+            let web_identity_token_provider = self.web_identity_builder.build();
+            let provider_chain = CredentialsProviderChain::first_try("Environment", env_provider)
+                .or_else("Profile", profile_provider)
+                .or_else("WebIdentityToken", web_identity_token_provider);
+            let cached_provider = self.credential_cache.load(provider_chain);
+            DefaultCredentialsChain(cached_provider.build())
+        }
+    }
+
+    #[cfg(test)]
+    mod test {
+
+        macro_rules! make_test {
+            ($name: ident) => {
+                #[traced_test]
+                #[tokio::test]
+                async fn $name() {
+                    crate::test_case::TestEnvironment::from_dir(concat!(
+                        "./test-data/default-provider-chain/",
+                        stringify!($name)
+                    ))
+                    .unwrap()
+                    .execute(|fs, env, conn| {
+                        crate::default_provider::credentials::Builder::default()
+                            .env(env)
+                            .fs(fs)
+                            .region(Region::from_static("us-east-1"))
+                            .connector(conn)
+                            .build()
+                    })
+                    .await
+                }
+            };
+        }
+
+        use aws_sdk_sts::Region;
+
+        use tracing_test::traced_test;
+
+        make_test!(prefer_environment);
+        make_test!(profile_static_keys);
+        make_test!(web_identity_token_env);
+        make_test!(web_identity_source_profile_no_env);
+        make_test!(web_identity_token_invalid_jwt);
+        make_test!(web_identity_token_source_profile);
+        make_test!(web_identity_token_profile);
+        make_test!(profile_overrides_web_identity);
+
+        /// Helper that uses `execute_and_update` instead of execute
+        ///
+        /// If you run this, it will add another HTTP traffic log which re-records the request
+        /// data
+        #[tokio::test]
+        #[ignore]
+        async fn update_test() {
+            crate::test_case::TestEnvironment::from_dir(concat!(
+                "./test-data/default-provider-chain/web_identity_token_source_profile",
+            ))
+            .unwrap()
+            .execute_and_update(|fs, env, conn| {
+                super::Builder::default()
+                    .env(env)
+                    .fs(fs)
+                    .region(Region::from_static("us-east-1"))
+                    .connector(conn)
+                    .build()
+            })
+            .await
+        }
     }
 }
