+++ aws/rust-runtime/aws-config/src/meta/credentials/mod.rs
@@ -14,7 +14,7 @@ pub use chain::CredentialsProviderChain;
 mod credential_fn;
 pub use credential_fn::async_provide_credentials_fn;
 
-mod lazy_caching;
+pub mod lazy_caching;
 pub use lazy_caching::LazyCachingCredentialsProvider;
 
 // pub mod credential_fn;
